% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCallList.R
\name{cleanName}
\alias{cleanName}
\title{Remove invalid names in multiple columns}
\usage{
cleanName(df, fields)
}
\arguments{
\item{df}{data frame containing at least two rows and two columns of names.}

\item{field}{character vector containing the field names of names in \code{df}.}
}
\value{
data frame
}
\description{
Invalid names in a specific record:
\enumerate{
  \item no names at all.
}
}
