% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_data_t1}
\alias{customer_data_t1}
\title{Customer List from 01/06/2011 to 30/11/2011}
\format{A tibble with 3,471 rows and 3 variables:
\describe{
  \item{CustomerID}{num. Customer number. Nominal, a 5-digit integral number uniquely assigned to each customer.}
  \item{visitDate}{POSIXct. last visit date of each customer during the period.}
  \item{Country}{chr. the latest (by invoice date during the period) country information of each customer}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Online+Retail}
}
\usage{
customer_data_t1
}
\description{
This is a customer data set which includes all customers from 01/06/2011 to 30/11/2011.
Each record represents each unique customer.
}
\details{
Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.
}
\keyword{datasets}
