% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRfmScore.R
\name{getRfmScore}
\alias{getRfmScore}
\title{Get RFM score}
\usage{
getRfmScore(df, r_score, f_score, m_score, weights, r_weight, f_weight,
  m_weight)
}
\arguments{
\item{df}{tibble.}

\item{r_score}{field name of recency score in \code{df}.}

\item{f_score}{field name of frequency score in \code{df}.}

\item{m_score}{field name of monetary score in \code{df}.}

\item{weights}{a numeric vector containing RFM weights, must have names.}

\item{r_weight}{name of recency weight in \code{weights}.}

\item{f_weight}{name of frequency weight in \code{weights}.}

\item{m_weight}{name of monetary weight in \code{weights}.}
}
\value{
an original tibble with an extra column of weighted RFM scores.
}
\description{
Get RFM score
}
