% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRfmTable.R
\name{getRfmTable}
\alias{getRfmTable}
\title{Get RFM values and scores}
\usage{
getRfmTable(customerData, transactionData, id, tData_visitDate,
  tData_revenue, cData_visitDate = NULL, recencyFloorDate = NULL,
  reportDate = NULL, bins = 5)
}
\arguments{
\item{customerData}{tibble, customer unique records, must contain primary key.}

\item{transactionData}{tibble, transaction records, must contain foreign key.}

\item{id}{key name, should be found in both \code{customerData} and \code{transactionData}.}

\item{tData_visitDate}{field name of visit date column in \code{transactionData}. This field must be in the class of \code{POSIXct}. Any records containing \code{NA} will be removed.}

\item{tData_revenue}{field name of revenue column in \code{transactionData}, any records containing \code{NA} will be removed.}

\item{cData_visitDate}{field name of last visit date in \code{customerData}. This field must be in the class of \code{POSIXct};
optional, default value is \code{NULL}. If using default value, \code{recencyFloorDate} will be applied to fill \code{tData_visitDate}
for the cases of "did not come back" (customers were only found in \code{customerData}, but could not be found in \code{transactionData}).
If not using default value, the specified field will be applied to fill \code{tData_visitDate}
for the cases of "did not come back".
If the specified field contains missing values, highly recommend to use \code{recencyFloorDate} instead.}

\item{recencyFloorDate}{recency floor date for "did not come back" customers;
optional, default value is the oldest date of \code{tData_visitDate}, must be in the class of \code{POSIXct}.}

\item{reportDate}{recency ceiling date; optional, default value is the newest date of \code{tData_visitDate},
must be in the class of \code{POSIXct}.}

\item{bins}{bins for RFM scores; optional, default value is 5.}
}
\value{
A tibble containing RFM values, percentile rank, scores and z-score.
}
\description{
Get RFM values and scores
}
