% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRfmWeights.R
\name{getRfmWeights}
\alias{getRfmWeights}
\title{Get RFM weights}
\usage{
getRfmWeights(df, formula, family = "binomial", show = TRUE)
}
\arguments{
\item{df}{data frame or tibble.}

\item{formula}{formula for the Generalized Linear Model (GLM).}

\item{family}{family for the GLM; optional, default value is \code{binomial}.}

\item{show}{show regression table and weights; optional, default value is \code{TRUE}.}
}
\value{
A list containing an atomic vector of weights and a GLM model.
}
\description{
Get RFM weights
}
