% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroupBars.R
\name{plotGroupBars}
\alias{plotGroupBars}
\title{Plot Bar Chart by Group}
\usage{
plotGroupBars(df, groupVar, measureVar, title, label, fun = mean,
  dp = 1, mean_line = TRUE, ...)
}
\arguments{
\item{df}{data frame or tibble.}

\item{groupVar}{field name for groupping in \code{df}.}

\item{measureVar}{field name of the measure in \code{df}.}

\item{title}{chart's title.}

\item{label}{label for \code{measureVar}.}

\item{fun}{aggregate function; optional, default value is \code{mean()}.}

\item{dp}{number of decimal places; optional, default value is 1.}

\item{mean_line}{show line of mean value; optional; default value is \code{TRUE}.}

\item{...}{extra arguments go to \code{geom_bar()}.}
}
\value{
a ggplot
}
\description{
Plot Bar Chart by Group
}
