% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatMap.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{Plot Heat Map of RFM}
\usage{
plotHeatMap(df, r_score, f_score, m_value, rfm_score, low = "cornsilk",
  high = "orange", size = 4, ...)
}
\arguments{
\item{df}{data frame or tibble.}

\item{r_score}{field name of recency score in \code{df}.}

\item{f_score}{field name of frequency score in \code{df}.}

\item{m_value}{field name of monetary value in \code{df}.}

\item{rfm_score}{field name of weighted RFM score in \code{df}.}

\item{low}{color for low value of \code{m_value}; default value is \code{cornsilk}.}

\item{high}{color for high value of \code{m_value}; default value is \code{orange}.}

\item{size}{the size of annotation; optional; default value is 4.}

\item{...}{extra arguments go to \code{geom_tile()}.}
}
\value{
a ggplot.
}
\description{
Plot Heat Map of RFM
}
