% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiGroupBars.R
\name{plotPeriodGroupBars}
\alias{plotPeriodGroupBars}
\title{Plot Bar Chart by Group Over Time}
\usage{
plotPeriodGroupBars(df, groupVar, measureVar, periodVar, sortFirst, title,
  label, legendTitle, fun = mean, dp = 1, colorVector = c("gold",
  "orange"), ...)
}
\arguments{
\item{df}{data frame or tibble.}

\item{groupVar}{field name for groupping in \code{df}.}

\item{measureVar}{field name of the measure in \code{df}.}

\item{periodVar}{field name of the period in \code{df}.}

\item{sortFirst}{value in \code{periodVar}. sort the bars with this value first.}

\item{title}{chart's title.}

\item{label}{label for \code{measureVar}.}

\item{fun}{aggregate function; optional, default value is \code{mean()}.}

\item{dp}{number of decimal places; optional, default value is 1.}

\item{colorVector}{character vector for setting the color of the bars; optional, default value is \code{c("gold", "orange")}.}

\item{...}{extra arguments go to \code{geom_bar()}.}

\item{lagendTitle}{title of the legend.}
}
\value{
a ggplot
}
\description{
Plot Bar Chart by Group Over Time
}
