% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCallList.R
\name{removeDup}
\alias{removeDup}
\title{Remove invalid phone numbers in multiple columns}
\usage{
removeDup(df, fields, details = "all")
}
\arguments{
\item{df}{data frame containing at least two rows and two columns of phone numbers.}

\item{details}{optional, default value is \code{all} which returns extra information. \code{onlyFields} omits such information.}

\item{field}{character vector containing the field names of phone numbers in \code{df}.}
}
\value{
data frame
}
\description{
Invalid phone numbers in a specific record:
\enumerate{
  \item all phone numbers that are not equal to 8 digits, including no phone numbers at all (all zero digit).
  \item all phone numbers that start from number zero or one.
  \item all phone numbers that duplicate between other records, excluding phone numbers duplicate within the same record.
}
}
