% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transaction_data_all}
\alias{transaction_data_all}
\title{Online Retail Data Set}
\format{A tibble with 541,909 rows and 8 variables:
\describe{
  \item{InvoiceNo}{chr. Invoice number. Nominal, a 6-digit integral number uniquely assigned to each transaction. If this code starts with letter 'c', it indicates a cancellation.}
  \item{StockCode}{chr. Product (item) code. Nominal, a 5-digit integral number uniquely assigned to each distinct product.}
  \item{Description}{chr. Product (item) name. Nominal.}
  \item{Quantity}{num. The quantities of each product (item) per transaction. Numeric.}
  \item{InvoiceDate}{POSIXct. Invoice Date and time. Numeric, the day and time when each transaction was generated.}
  \item{UnitPrice}{num. Unit price. Numeric, Product price per unit in sterling.}
  \item{CustomerID}{num. Customer number. Nominal, a 5-digit integral number uniquely assigned to each customer.}
  \item{Country}{chr. Country name. Nominal, the name of the country where each customer resides.}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Online+Retail}
}
\usage{
transaction_data_all
}
\description{
This is a transnational data set which contains all the transactions
occurring between 01/12/2010 and 09/12/2011 for a UK-based and registered
non-store online retail.
}
\details{
Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.
}
\keyword{datasets}
