% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transaction_data_t1}
\alias{transaction_data_t1}
\title{Transaction records from 01/06/2011 to 30/11/2011}
\format{A tibble with 235,130 rows and 9 variables:
\describe{
  \item{InvoiceNo}{chr. Invoice number. Nominal, a 6-digit integral number uniquely assigned to each transaction. If this code starts with letter 'c', it indicates a cancellation.}
  \item{StockCode}{chr. Product (item) code. Nominal, a 5-digit integral number uniquely assigned to each distinct product.}
  \item{Description}{chr. Product (item) name. Nominal.}
  \item{Quantity}{num. The quantities of each product (item) per transaction. Numeric.}
  \item{InvoiceDate}{POSIXct. Invoice Date and time. Numeric, the day and time when each transaction was generated.}
  \item{UnitPrice}{num. Unit price. Numeric, Product price per unit in sterling.}
  \item{CustomerID}{num. Customer number. Nominal, a 5-digit integral number uniquely assigned to each customer.}
  \item{Country}{chr. Country name. Nominal, the name of the country where each customer resides.}
  \item{revenue}{num. UnitPrice x Quantity}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Online+Retail}
}
\usage{
transaction_data_t1
}
\description{
This is a transactional data set which includes all customers' transaction
from 01/06/2011 to 30/11/2011.
Each record represents each transaction of customers.
}
\details{
Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.
}
\keyword{datasets}
